Move code into src package
