+++ main.py
@@ -1,6 +1,6 @@
 """Main entry point for starting the Telegram Bot."""
 
-from core.cli import cli
+from src.core.cli import cli
 
 
 if __name__ == "__main__":
