+++ src/core/cli.py
@@ -2,11 +2,11 @@ import argparse
 import logging
 import sys
 
-from config import RUN_MODE, validate_config
-from utils.environment import check_environment
-from utils.logger import setup_logging
+from src.config import RUN_MODE, validate_config
+from src.utils.environment import check_environment
+from src.utils.logger import setup_logging
 
-from core.runner import run_telegram_bot
+from src.core.runner import run_telegram_bot
 
 
 def create_parser() -> argparse.ArgumentParser:
