+++ src/core/runner.py
@@ -5,14 +5,20 @@ and startup mode (polling or webhook). Provides entry points for bot execution
 and integrates logging, command registration, and graceful exception handling.
 """
 
-from config import BOT_TOKEN, RUN_MODE, WEBHOOK_LISTEN, WEBHOOK_PORT, WEBHOOK_URL
-from handlers.fallback import unknown_command
-from handlers_loader import register_handlers
+from src.config import (
+    BOT_TOKEN,
+    RUN_MODE,
+    WEBHOOK_LISTEN,
+    WEBHOOK_PORT,
+    WEBHOOK_URL,
+)
+from src.handlers.fallback import unknown_command
+from src.handlers_loader import register_handlers
 from telegram.ext import Application, ApplicationBuilder, MessageHandler, filters
-from utils.commands import make_set_commands
-from utils.logger import logger
+from src.utils.commands import make_set_commands
+from src.utils.logger import logger
 
-from core.error_handler import handle_error
+from src.core.error_handler import handle_error
 
 
 def create_application() -> Application:
