+++ src/handlers/fallback.py
@@ -8,7 +8,7 @@ import logging
 
 from telegram import Update
 from telegram.ext import ContextTypes
-from utils.commands import command
+from src.utils.commands import command
 
 logger = logging.getLogger("bot_bot")
